requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
oracledb==2.2.1
python-dotenv==1.0.1
